factor out clustered server
